backend/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
customtkinter==5.2.0
darkdetect==0.8.0
Django==4.2.4
django-cors-headers==4.2.0
django-jsonform==2.19.1
djangorestframework==3.14.0
idna==3.4
Python-IO==0.3
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
psycopg2-binary==2.9.9